# atm-service/src/main/java/com/egc/atmservice/domain/receipt/Receipt.java
package com.egc.atmservice.domain.receipt;

import lombok.Data;

import java.math.BigDecimal;

@Data
public class Receipt {

    private String cardNumber;

    private BigDecimal balance;

    private BigDecimal transactionAmount;

    private TransactionType transactionType;
}